refactor(test): rename and restructure test files for organization
- Rename test files to reflect their actual test scope
  (test_async_client -> test_jsonplaceholder_client_async, etc.)
- Add new test files for exceptions, response_parsing, base classes
- Remove redundant old test files replaced by restructuring
