+++ tests/unit/test_http_helpers.py
@@ -1,28 +1,14 @@
-"""Shared tests for module-level utility functions in api_client.py (non-async-specific).
+"""HTTP helper tests for utils.http_helpers."""
 
-These tests verify exception hierarchy, _safe_parse_json(), _map_request_error(),
-_resolve_client_config(), and _classify_error()
-which are module-level utility functions independent of AsyncAPIClient/SyncAPIClient.
-
-Consolidated from test_async_client_error_handling.py and
-test_sync_client_error_handling.py to eliminate duplication.
-"""
-
-import json
-from collections.abc import Callable
 from unittest.mock import Mock, patch
 
 import httpx
 import pytest
-from pydantic import BaseModel, Field
 
 from tests.constants import INVALID_BASE_URLS
 from utils.exceptions import (
     APIClientError,
     APIConnectionError,
-    APIHTTPError,
-    APIJSONDecodeError,
-    APIRetryError,
     APITimeoutError,
 )
 from utils.http_helpers import (
@@ -37,203 +23,8 @@ from utils.http_helpers import (
 from utils.http_helpers import (
     resolve_client_config as _resolve_client_config,
 )
-from utils.jsonplaceholder_base_async import AsyncAPIClient
-from utils.jsonplaceholder_base_sync import SyncAPIClient
-from utils.response_parsing import (
-    parse_response_model as _parse_response_model,
-)
-from utils.response_parsing import (
-    parse_response_model_list as _parse_response_model_list,
-)
-from utils.response_parsing import (
-    safe_parse_json as _safe_parse_json,
-)
 
-# Module-level marker: All tests in this file are unit tests
 pytestmark = pytest.mark.unit
-
-
-# =============================================================================
-# 型定義（Type Alias）
-# =============================================================================
-
-# 引数に Optional[object] を受け取り、戻り値として Mock を返す
-MockResponseFactory = Callable[[object | None], Mock]
-
-
-@pytest.fixture()
-def mock_response_factory() -> MockResponseFactory:
-    """テスト用 Mock(spec=httpx.Response) を生成する factory fixture
-
-    テスト間で mock 生成ロジックを集約し、httpx.Response 仕様変更時の
-    修正を1箇所に限定する。
-
-    Returns:
-        payload（省略可）を受け取り、json.return_value を設定した
-        Mock(spec=httpx.Response) を返す callable
-    """
-
-    def _factory(payload: object | None = None) -> Mock:
-        response = Mock(spec=httpx.Response)
-        if payload is not None:
-            response.json.return_value = payload
-        return response
-
-    return _factory
-
-
-# =============================================================================
-# Pydantic Parsing Tests（_parse_response_model / _parse_response_model_list の検証）
-# =============================================================================
-
-
-class DummyModel(BaseModel):
-    """テスト用のシンプルなPydanticモデル"""
-
-    id: int = Field(..., ge=1)
-    name: str
-
-
-def test_parse_response_model_success(mock_response_factory: MockResponseFactory) -> None:
-    """_parse_response_model: 正常系（dict -> model）"""
-    mock_response = mock_response_factory({"id": 1, "name": "test"})
-
-    result = _parse_response_model(mock_response, DummyModel)
-
-    assert isinstance(result, DummyModel)
-    assert result.id == 1
-    assert result.name == "test"
-
-
-def test_parse_response_model_invalid_type(mock_response_factory: MockResponseFactory) -> None:
-    """_parse_response_model: 異常系（配列が返ってきた場合）"""
-    mock_response = mock_response_factory([{"id": 1, "name": "test"}])
-
-    with pytest.raises(APIJSONDecodeError, match="Expected object JSON for DummyModel, got list"):
-        _parse_response_model(mock_response, DummyModel)
-
-
-def test_parse_response_model_validation_error(mock_response_factory: MockResponseFactory) -> None:
-    """_parse_response_model: 異常系（バリデーションエラー）"""
-    mock_response = mock_response_factory({"id": 0, "name": "test"})  # id < 1
-
-    with pytest.raises(APIJSONDecodeError, match="Invalid DummyModel response schema"):
-        _parse_response_model(mock_response, DummyModel)
-
-
-def test_parse_response_model_list_success(mock_response_factory: MockResponseFactory) -> None:
-    """_parse_response_model_list: 正常系（list -> list[model]）"""
-    mock_response = mock_response_factory([{"id": 1, "name": "test1"}, {"id": 2, "name": "test2"}])
-
-    result = _parse_response_model_list(mock_response, DummyModel)
-
-    assert isinstance(result, list)
-    assert len(result) == 2
-    assert all(isinstance(item, DummyModel) for item in result)
-    assert result[0].id == 1
-    assert result[1].name == "test2"
-
-
-def test_parse_response_model_list_invalid_type(mock_response_factory: MockResponseFactory) -> None:
-    """_parse_response_model_list: 異常系（オブジェクトが返ってきた場合）"""
-    mock_response = mock_response_factory({"id": 1, "name": "test"})
-
-    with pytest.raises(APIJSONDecodeError, match="Expected array JSON for DummyModel, got dict"):
-        _parse_response_model_list(mock_response, DummyModel)
-
-
-def test_parse_response_model_list_validation_error(
-    mock_response_factory: MockResponseFactory,
-) -> None:
-    """_parse_response_model_list: 異常系（要素にバリデーションエラーがある場合）
-
-    エラーメッセージに失敗要素の index（loc 先頭）が含まれ、配列内の
-    どの要素が原因か診断できることを検証する（TypeAdapter による index 付与）。
-    """
-    mock_response = mock_response_factory([{"id": 1, "name": "test1"}, {"id": -1, "name": "test2"}])
-
-    with pytest.raises(APIJSONDecodeError, match="Invalid DummyModel response schema") as exc_info:
-        _parse_response_model_list(mock_response, DummyModel)
-
-    # index=1（2番目の要素）が失敗したことが loc 先頭に表れる
-    assert "1.id" in str(exc_info.value)
-
-
-# =============================================================================
-# Exception Tests（例外クラスの検証）
-# =============================================================================
-
-
-def test_exception_hierarchy() -> None:
-    """例外クラスの継承関係確認"""
-    assert issubclass(APIConnectionError, APIClientError)
-    assert issubclass(APITimeoutError, APIClientError)
-    assert issubclass(APIHTTPError, APIClientError)
-    assert issubclass(APIRetryError, APIClientError)
-    assert issubclass(APIJSONDecodeError, APIClientError)
-    assert issubclass(APIClientError, Exception)
-
-
-def test_http_error_status_preservation(
-    mock_response_factory: MockResponseFactory,
-) -> None:
-    """APIHTTPError がステータスコードを保持することを確認"""
-    mock_response = mock_response_factory()
-    mock_response.status_code = 404
-
-    error = APIHTTPError("Not Found", status_code=404, response=mock_response)
-
-    assert error.status_code == 404
-    assert error.response == mock_response
-    assert str(error) == "Not Found"
-
-
-def test_retry_error_message() -> None:
-    """APIRetryError のメッセージ確認"""
-    error = APIRetryError("Max retries exceeded")
-    assert str(error) == "Max retries exceeded"
-
-
-# =============================================================================
-# JSON Parsing Tests（_safe_parse_json の検証）
-# =============================================================================
-
-
-def test_safe_parse_json_invalid_json(
-    mock_response_factory: MockResponseFactory,
-) -> None:
-    """不正なJSONでAPIJSONDecodeErrorが発生（エラーパス）"""
-    mock_response = mock_response_factory()
-    mock_response.json.side_effect = json.JSONDecodeError("Invalid JSON", "doc", 0)
-
-    with pytest.raises(APIJSONDecodeError) as exc_info:
-        _safe_parse_json(mock_response)
-
-    assert "Failed to parse JSON" in str(exc_info.value)
-    assert "Invalid JSON" in str(exc_info.value)  # str(e) の診断情報を保持
-    assert exc_info.value.response == mock_response
-
-
-def test_api_json_decode_error_init(mock_response_factory: MockResponseFactory) -> None:
-    """APIJSONDecodeErrorのコンストラクタテスト"""
-    mock_response = mock_response_factory()
-    error = APIJSONDecodeError("Parse error", response=mock_response)
-
-    assert str(error) == "Parse error"
-    assert error.response == mock_response
-
-
-def test_api_json_decode_error_without_response() -> None:
-    """APIJSONDecodeError: responseなしでも動作"""
-    error = APIJSONDecodeError("Parse error")
-
-    assert str(error) == "Parse error"
-    assert error.response is None
-
-
-# =============================================================================
-# Request Error Mapping Tests（_map_request_error の検証）
-# =============================================================================
 
 
 def test_map_request_error_too_many_redirects() -> None:
@@ -301,11 +92,6 @@ def test_map_request_error_network_error() -> None:
     # セキュリティ: str(e) が含まれないこと（機密情報漏洩防止）
     assert "proxy.internal.example.com" not in str(result)
     assert "NetworkError" in str(result)  # type(e).__name__ が含まれること
-
-
-# =============================================================================
-# Classify Error Tests（_classify_error の検証）
-# =============================================================================
 
 
 def test_classify_error_non_retryable_logs_error() -> None:
@@ -395,12 +181,14 @@ def test_classify_error_retryable_network_error() -> None:
     assert "error" not in call_kwargs[1]
 
 
-# =============================================================================
-# Resolve Client Config Tests（_resolve_client_config の検証）
-# =============================================================================
-
-
 class MockAPISettings:
+    """settings フォールバック検証用の手書き fake（本物の APIConfig は使えない）。
+
+        Note: base_url の "settings.example.com" は ALLOWED_DOMAINS allowlist 外のため、
+    `APIConfig(base_url=...)` は SSRF Prevention バリデータで ValidationError を送出する。
+        この専用ホストは settings 由来の値であることの目印として assert されている。
+    """
+
     def __init__(self) -> None:
         self.base_url = "https://settings.example.com"
         self.timeout = 30.0
@@ -536,49 +324,6 @@ def test_resolve_client_config_zero_retry_delay_not_overridden(mock_settings: Mo
             "https://example.com", None, None, 0.0, None
         )
     assert retry_delay == 0.0
-
-
-# =============================================================================
-# Client Init Tests（クライアント初期化の検証）
-# =============================================================================
-
-
-def test_sync_client_headers_empty_dict_preserves_defaults(mock_base_url: str) -> None:
-    """SyncAPIClient: headers={} でデフォルトヘッダーが保持される
-
-    `if headers is not None:` の設計を保証するテスト。
-    空辞書を渡しても update({}) は no-op なので、デフォルトヘッダーは変わらない。
-    """
-    with SyncAPIClient(base_url=mock_base_url, headers={}) as client:
-        # デフォルトヘッダーは3つ: User-Agent, Accept, Content-Type
-        assert set(client.default_headers.keys()) == {
-            "User-Agent",
-            "Accept",
-            "Content-Type",
-        }
-        # Noneではなく、正しく設定されている
-        assert client.default_headers["User-Agent"]
-        assert client.default_headers["Accept"] == "application/json"
-        assert client.default_headers["Content-Type"] == "application/json"
-
-
-async def test_async_client_headers_empty_dict_preserves_defaults(mock_base_url: str) -> None:
-    """AsyncAPIClient: headers={} でデフォルトヘッダーが保持される
-
-    `if headers is not None:` の設計を保証するテスト。
-    空辞書を渡しても update({}) は no-op なので、デフォルトヘッダーは変わらない。
-    """
-    async with AsyncAPIClient(base_url=mock_base_url, headers={}) as client:
-        # デフォルトヘッダーは3つ: User-Agent, Accept, Content-Type
-        assert set(client.default_headers.keys()) == {
-            "User-Agent",
-            "Accept",
-            "Content-Type",
-        }
-        # Noneではなく、正しく設定されている
-        assert client.default_headers["User-Agent"]
-        assert client.default_headers["Accept"] == "application/json"
-        assert client.default_headers["Content-Type"] == "application/json"
 
 
 class TestLogErrorWithStderrFallback:
